fix(modal-divider): use dataAttr instead of variant
