+++ packages/admin-ui/src/divider/index.ts
@@ -1,0 +1,1 @@
+export * from './divider'
