+++ packages/admin-ui/src/index.ts
@@ -39,6 +39,7 @@ export * from './skeleton'
 export * from './pagination'
 export * from './data-view'
 export * from './modal'
+export * from './divider'
 
 export * from './components/Label'
 export * from './components/CollapsibleGroup'
@@ -48,7 +49,6 @@ export * from './components/Paragraph'
 export * from './components/List'
 export * from './components/Heading'
 export * from './components/Text'
-export * from './components/Divider'
 export * from './components/Dropdown'
 export * from './components/Spinner'
 export * from './components/Sidebar'
